Added more file structure, began abstracting into structure
